feat: moves software updates to interface settings page
