+++ src/components/widgets/stats/PrinterStatsCard.vue
@@ -32,7 +32,7 @@
       </div>
 
       <v-row v-if="supportsHistoryPlugin">
-        <v-col :cols="breakpoint">
+        <v-col cols="6">
           <v-card outlined class="px-2 py-1 text-center stat-square">
             <div class="grey--text text--darken-2">{{ $t('app.general.label.total_jobs') }}</div>
             <div class="grey--text focus--text">{{ rollup.total_jobs }}</div>
@@ -40,7 +40,7 @@
             <div class="grey--text focus--text">{{ $filters.formatCounterTime(rollup.longest_job) }}</div>
           </v-card>
         </v-col>
-        <v-col :cols="breakpoint">
+        <v-col cols="6">
           <v-card outlined class="px-2 py-1 text-center stat-square">
             <div class="grey--text text--darken-2">{{ $t('app.general.label.total_time') }}</div>
             <div class="grey--text focus--text">{{ $filters.formatCounterTime(rollup.total_time) }}</div>
@@ -48,7 +48,7 @@
             <div class="grey--text focus--text">{{ $filters.formatCounterTime(rollup.total_avg) }}</div>
           </v-card>
         </v-col>
-        <v-col :cols="breakpoint">
+        <v-col cols="6">
           <v-card outlined class="px-2 py-1 text-center stat-square">
             <div class="grey--text text--darken-2">{{ $t('app.general.label.total_print_time') }}</div>
             <div class="grey--text focus--text">{{ $filters.formatCounterTime(rollup.total_print_time) }}</div>
@@ -56,7 +56,7 @@
             <div class="grey--text focus--text">{{ $filters.formatCounterTime(rollup.print_avg) }}</div>
           </v-card>
         </v-col>
-        <v-col :cols="breakpoint">
+        <v-col cols="6">
           <v-card outlined class="px-2 py-1 text-center stat-square">
             <div class="grey--text text--darken-2">{{ $t('app.general.label.total_filament') }}</div>
             <div class="grey--text focus--text">{{ $filters.getReadableLengthString(rollup.total_filament_used) }}</div>
@@ -79,13 +79,6 @@ import JobHistory from '@/components/widgets/history/JobHistory.vue'
   }
 })
 export default class PrinterStatsCard extends Vue {
-  get breakpoint () {
-    if (this.$vuetify.breakpoint.mdAndDown) {
-      return 6
-    }
-    return 3
-  }
-
   get rollup () {
     return this.$store.getters['history/getRollUp']
   }
